# openapi.yaml
openapi: 3.0.3
info:
  title: MVTMS REST API
  description: |
    OpenAPI specification for the MVTMS REST API v1. Authentication is via
    a signed Bearer token (`Authorization: Bearer <token>`) or an existing
    session cookie. The token endpoints are public; all other `/api/v1/*`
    routes require a valid principal plus Casbin RBAC permissions.
  version: "1.0.0"
  contact:
    name: MVTMS Support
    url: https://avandab.com
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/errors:
    get:
      tags: [Errors]
      summary: List error report groups
      description: Requires `errors:read` permission. Returns deduplicated error groups (fingerprint + tenant scoped).
      parameters:
        - name: severity
          in: query
          schema: { type: string, enum: [CRITICAL, HIGH, MEDIUM, LOW] }
        - name: fingerprint
          in: query
          schema: { type: string }
        - name: from
          in: query
          description: RFC3339 or YYYY-MM-DD, filters on last_seen.
          schema: { type: string }
        - name: to
          in: query
          description: RFC3339 or YYYY-MM-DD, filters on last_seen.
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: Error groups with total count.
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items: { $ref: '#/components/schemas/ErrorReport' }
                  total: { type: integer }
                  limit: { type: string }
                  offset: { type: string }
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/errors/{fingerprint}:
    get:
      tags: [Errors]
      summary: Get one error report group
      description: Requires `errors:read` permission. Powers the /ops/errors detail drawer.
      parameters:
        - name: fingerprint
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Full error group detail (message, stack trace, metadata, occurrence stats).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorReport' }
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/invoices:
    get:
      tags: [Invoices]
      summary: List invoices
      description: Requires `invoices:read` permission.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SearchParam'
        - $ref: '#/components/parameters/StatusParam'
      responses:
        '200':
          description: Paginated list of invoices.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Invoices]
      summary: Generate an invoice
      description: Requires `invoices:create` permission.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateInvoiceRequest'
      responses:
        '201':
          description: Invoice generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IDResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/invoices/{id}:
    get:
      tags: [Invoices]
      summary: Get an invoice
      description: Requires `invoices:read` permission.
      parameters:
        - $ref: '#/components/parameters/IDParam'
      responses:
        '200':
          description: Invoice details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/payments:
    get:
      tags: [Payments]
      summary: List payments
      description: Requires `payments:read` permission.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: method
          in: query
          schema:
            type: string
          description: Filter by payment method.
      responses:
        '200':
          description: Paginated list of payments.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Payments]
      summary: Record a payment
      description: Requires `payments:create` permission.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordPaymentRequest'
      responses:
        '201':
          description: Payment recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IDResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/payments/{id}:
    get:
      tags: [Payments]
      summary: Get a payment
      description: Requires `payments:read` permission.
      parameters:
        - $ref: '#/components/parameters/IDParam'
      responses:
        '200':
          description: Payment details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Signed API token returned by `POST /api/v1/auth/token`.
  parameters:
    IDParam:
      name: id
      in: path
      required: true
      schema:
        type: string
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
      description: Page number (1-based).
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 10
      description: Items per page.
    SearchParam:
      name: search
      in: query
      schema:
        type: string
      description: Free-text search filter.
    StatusParam:
      name: status
      in: query
      schema:
        type: string
      description: Status filter.
  schemas:
    ErrorReport:
      type: object
      properties:
        id: { type: string }
        fingerprint: { type: string }
        timestamp: { type: string, format: date-time, description: Last occurrence (last_seen). }
        request_id: { type: string }
        user_id: { type: string }
        tenant_id: { type: string }
        url: { type: string }
        method: { type: string }
        status_code: { type: integer }
        stack_trace: { type: string }
        message: { type: string }
        severity: { type: string, enum: [CRITICAL, HIGH, MEDIUM, LOW] }
        environment: { type: string }
        app_version: { type: string }
        occurrences: { type: integer }
        first_seen: { type: string, format: date-time }
        metadata:
          type: object
          additionalProperties: true
          description: Client capture payload (breadcrumbs, viewport, source).
...
    IDResponse:
      type: object
      properties:
        id:
          type: string
...
    GenerateInvoiceRequest:
      type: object
      required: [booking_id, customer_id]
      properties:
        booking_id:
          type: string
        customer_id:
          type: string
        trip_id:
          type: string
          nullable: true
        subtotal:
          type: number
          minimum: 0
        tax:
          type: number
          minimum: 0
        discount:
          type: number
          minimum: 0
        total:
          type: number
          minimum: 0
    Invoice:
      type: object
      properties:
        id:
          type: string
        invoice_number:
          type: string
        booking_id:
          type: string
        booking_number:
          type: string
        customer_id:
          type: string
        customer_name:
          type: string
        customer_company:
          type: string
        trip_id:
          type: string
          nullable: true
        trip_number:
          type: string
        subtotal:
          type: number
        tax:
          type: number
        discount:
          type: number
        total:
          type: number
        payment_status:
          type: string
          enum: [pending, paid, partially_paid]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    InvoiceList:
      type: object
      properties:
        invoices:
          type: array
          items:
            $ref: '#/components/schemas/Invoice'
        total:
          type: integer
          format: int64
    RecordPaymentRequest:
      type: object
      required: [invoice_id, payment_date, amount, method]
      properties:
        invoice_id:
          type: string
        payment_date:
          type: string
          format: date-time
        amount:
          type: number
          exclusiveMinimum: 0
        method:
          type: string
          enum: [cash, upi, bank_transfer, cheque, razorpay]
        reference:
          type: string
          nullable: true
        remarks:
          type: string
          nullable: true
    Payment:
      type: object
      properties:
        id:
          type: string
        invoice_id:
          type: string
        invoice_number:
          type: string
        payment_date:
          type: string
          format: date-time
        amount:
          type: number
        method:
          type: string
          enum: [cash, upi, bank_transfer, cheque, razorpay]
        reference:
          type: string
          nullable: true
        remarks:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PaymentList:
      type: object
      properties:
        payments:
          type: array
          items:
            $ref: '#/components/schemas/Payment'
        total:
          type: integer
          format: int64
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
  responses:
    BadRequest:
      description: Bad request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required or invalid credentials.
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient RBAC permissions.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'